much

+++ server/src/models/jump.rs
@@ -1,9 +1,9 @@
-use serde::Serialize;
+use serde::{Serialize, Deserialize};
 
-#[derive(Serialize)]
+#[derive(Serialize, Deserialize, sqlx::FromRow)]
 pub struct Jump {
     pub jump_participant_id: i32,
     pub jump_round_id: i32,
-    pub jump_score: i32,
-    pub jump_distance: i32,
+    pub jump_score: f64,
+    pub jump_distance: f64,
 }
